feat(story-17-1): implement State Ownership Restructure
- Move uiReducer/userReducer from features/quiz/reducers/ to shared/store/ as .prelude.ts
- Remove listReducer/vocabLists from rootReducer.ts — vocabulary owns its own state
- Create shared/store/index.ts barrel export
- Update quiz barrel exports and reducers index
- Move test files to shared/store/__tests__/
- Remove vocabLists from 4 test mock states
- Delete old uiReducer.ts, userReducer.ts, and their tests from quiz feature
- Update BR/IML docs with status and completion details

AC: 12/12 complete. Tests: 30/30 files, 255/255 passed.
