+++ apps/frontend/src/features/quiz/reducers/index.ts
@@ -1,4 +1,2 @@
 export * from "./progressReducer";
 export * from "./rootReducer";
-export * from "./uiReducer";
-export * from "./userReducer";
