+++ apps/frontend/src/features/vocabulary/components/__tests__/Sidebar.test.tsx
@@ -50,7 +50,6 @@ describe("Sidebar", () => {
         isLoading: false,
         error: "",
       },
-      vocabLists: { itemsById: {}, itemIds: [] },
     };
     render(
       <ProgressStateContext.Provider value={mockState}>
@@ -84,7 +83,6 @@ describe("Sidebar", () => {
         isLoading: false,
         error: "",
       },
-      vocabLists: { itemsById: {}, itemIds: [] },
     };
     render(
       <ProgressStateContext.Provider value={mockState2}>
