+++ apps/frontend/src/shared/store/__tests__/uiStore.prelude.test.ts
@@ -1,5 +1,11 @@
-import { uiReducer, uiInitialState } from "../uiReducer";
-import type { UiAction } from "../uiReducer";
+/**
+ * uiStore.prelude.test.ts
+ *
+ * Tests for uiStore.prelude (moved from features/quiz/reducers/__tests__/uiReducer.test.ts, Story 17.1).
+ */
+
+import { uiReducer, uiInitialState } from "../uiStore.prelude";
+import type { UiAction } from "../uiStore.prelude";
 
 describe("uiReducer", () => {
   it("toggles loading", () => {
