+++ apps/frontend/src/shared/store/index.ts
@@ -1,0 +1,4 @@
+export { uiReducer, uiInitialState } from "./uiStore.prelude";
+export type { UiAction, UiState } from "./uiStore.prelude";
+export { userReducer, userInitialState } from "./userStore.prelude";
+export type { UserAction, UserState } from "./userStore.prelude";
